% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_mde.R
\name{dist_mde}
\alias{dist_mde}
\title{Construct a MDE-Family}
\usage{
dist_mde(n, m)
}
\arguments{
\item{n}{Number of dirac components, starting with a point mass at 0.}

\item{m}{Number of erlang components, translated by \code{n - 0.5}.}
}
\value{
A \code{MixtureDistribution} containing \code{n} \code{DiracDistribution}s at 0 .. n - 1 and a \code{TranslatedDistribution}
with offset \code{n - 0.5} of an \code{ErlangMixtureDistribution} with \code{m} shapes.
}
\description{
Constructs a MDE-Family distribution with fixed number of components.
}
\examples{
params <- list(
  dists = list(
    list(),
    list(dist = list(scale = 1.0, probs = list(0.5, 0.3, 0.2), shapes = list(1L, 2L, 3L)))
  ),
  probs = list(0.1, 0.9)
)
dist <- dist_mde(1, 3)
x <- dist$sample(100, with_params = params)
d_emp <- dist_empirical(x)

plot_distributions(
  empirical = d_emp,
  theoretical = dist,
  with_params = list(
    theoretical = params
  ),
  .x = seq(1e-4, 5, length.out = 100)
)

}
\seealso{
Other Distributions: 
\code{\link{Distribution}},
\code{\link{dist_bdegp}()},
\code{\link{dist_beta}()},
\code{\link{dist_binomial}()},
\code{\link{dist_blended}()},
\code{\link{dist_dirac}()},
\code{\link{dist_discrete}()},
\code{\link{dist_empirical}()},
\code{\link{dist_erlangmix}()},
\code{\link{dist_exponential}()},
\code{\link{dist_gamma}()},
\code{\link{dist_genpareto}()},
\code{\link{dist_lognormal}()},
\code{\link{dist_mixture}()},
\code{\link{dist_negbinomial}()},
\code{\link{dist_normal}()},
\code{\link{dist_pareto}()},
\code{\link{dist_poisson}()},
\code{\link{dist_translate}()},
\code{\link{dist_trunc}()},
\code{\link{dist_uniform}()},
\code{\link{dist_weibull}()}
}
\concept{Distributions}
