% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_union.R
\name{interval_union}
\alias{interval_union}
\title{Compute a union of intervals}
\usage{
interval_union(...)
}
\arguments{
\item{...}{Possibly list-nested intervals, interval_unions or numerics}
}
\value{
An \code{interval_union}
}
\description{
Computes a union of a (possibly nested) list of intervals and points, simplifying as much as possible.
}
\details{
An \code{interval_union} is a matrix with columns "lowest", "highest", "include_lowest", "include_highest" and
"integer", efficiently representing a list of intervals with ranges (lowest, highest).
Use \code{as.list(union)} to construct a list of \code{Interval}s representing the same union if needed.
}
\examples{
interval_union(interval(0, 1), interval(1, 3))
interval_union(interval(0, 1, closed = TRUE), interval(1, 3))
interval_union(interval(0, 1, closed = TRUE), interval(2, 3, closed = TRUE))
interval_union(interval(0, 1, closed = TRUE, integer = TRUE), interval(2, 3, closed = TRUE, integer = TRUE))
interval_union(interval(0, 1, closed = TRUE), interval(0.5, 2, include_lowest = TRUE))
interval_union(
  interval(0, 1, closed = TRUE, integer = TRUE),
  interval(0.5, 2, include_lowest = TRUE, integer = TRUE),
  2.5, 3.5, -1.0
)

# Compute a list of atomic intervals necessary to represent the interval union
as.list(
  interval_union(interval(0, 1, closed = TRUE, integer = TRUE), interval(0.5, 2, closed = TRUE, integer = TRUE))
)
}
