% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_class.R
\name{is.Distribution}
\alias{is.Distribution}
\title{Test if object is a Distribution}
\usage{
is.Distribution(object)
}
\arguments{
\item{object}{An R object.}
}
\value{
\code{TRUE} if \code{object} is a Distribution, \code{FALSE} otherwise.
}
\description{
Test if object is a Distribution
}
\examples{
is.Distribution(dist_dirac())

}
